requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
anyio==3.6.2
arrow==1.2.3
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.2.0
backoff==2.2.1
binaryornot==0.4.4
certifi==2022.12.7
chardet==5.1.0
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
cookiecutter==1.7.3
distlib==0.3.6
Django==4.1.5
dnspython==2.3.0
email-validator==1.3.1
fastapi==0.89.1
filelock==3.9.0
frozenlist==1.3.3
greenlet==2.0.1
h11==0.14.0
httpcore==0.16.3
httptools==0.5.0
httpx==0.23.3
idna==3.4
iniconfig==2.0.0
Jinja2==3.1.2
jinja2-time==0.2.0
manage-fastapi==1.1.1
MarkupSafe==2.1.2
multidict==6.0.4
openai==0.26.1
openai-async==0.0.2
packaging==23.0
platformdirs==2.6.2
pluggy==1.0.0
poyo==0.5.0
prompt-toolkit==3.0.36
psycopg2==2.9.5
pydantic==1.10.4
pytest==7.2.1
python-dateutil==2.8.2
python-dotenv==0.21.0
python-multipart==0.0.5
python-slugify==7.0.0
pytz==2022.7.1
PyYAML==6.0
questionary==1.10.0
requests==2.28.2
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.3.24
sqlparse==0.4.3
starlette==0.22.0
tenacity==8.1.0
text-unidecode==1.3
tqdm==4.64.1
typer==0.4.2
typing_extensions==4.4.0
tzdata==2022.7
urllib3==1.26.14
uvicorn==0.20.0
virtualenv==20.17.1
watchfiles==0.18.1
wcwidth==0.2.6
websockets==10.4
yarl==1.8.2
